+++ toolium/utils/dataset.py
@@ -56,41 +56,44 @@ def replace_param(param, language='es', infer_param_type=True):
     """
     Apply transformations to the given param based on specific patterns.
     Available replacements:
-        [STRING_WITH_LENGTH_XX] Generates a fixed length string
-        [INTEGER_WITH_LENGTH_XX] Generates a fixed length integer
-        [STRING_ARRAY_WITH_LENGTH_XX] Generates a fixed length array of strings
-        [INTEGER_ARRAY_WITH_LENGTH_XX] Generates a fixed length array of integers
-        [JSON_WITH_LENGTH_XX] Generates a fixed length JSON
-        [MISSING_PARAM] Generates a None object
-        [NULL] Generates a None object
-        [TRUE] Generates a boolean True
-        [FALSE] Generates a boolean False
-        [EMPTY] Generates an empty string
-        [B] Generates a blank space
-        [UUID] Generates a v4 UUID
-        [RANDOM] Generates a random value
-        [RANDOM_PHONE_NUMBER] Generates a random phone number for language and country configured in dataset.language
-            and dataset.country
-        [TIMESTAMP] Generates a timestamp from the current time
-        [DATETIME] Generates a datetime from the current time
-        [NOW] Similar to DATETIME without milliseconds; the format depends on the language
-        [NOW(%Y-%m-%dT%H:%M:%SZ)] Same as NOW but using an specific format by the python strftime function of the
-            datetime module
-        [NOW + 2 DAYS] Similar to NOW but two days later
-        [NOW - 1 MINUTES] Similar to NOW but one minute earlier
-        [NOW(%Y-%m-%dT%H:%M:%SZ) - 7 DAYS] Similar to NOW but seven days before and with the indicated format
-        [TODAY] Similar to NOW without time; the format depends on the language
-        [TODAY + 2 DAYS] Similar to NOW, but two days later
-        [ROUND:xxxx::y] Generates a string from a float number (xxxx) with the indicated number of decimals (y)
-        [STR:xxxx] Cast xxxx to a string
-        [INT:xxxx] Cast xxxx to an int
-        [FLOAT:xxxx] Cast xxxx to a float
-        [LIST:xxxx] Cast xxxx to a list
-        [DICT:xxxx] Cast xxxx to a dict
-        [UPPER:xxxx] Converts xxxx to upper case
-        [LOWER:xxxx] Converts xxxx to lower case
-        [REPLACE:xxxxx::yy::zz] Replace elements in string. Example: [REPLACE:[CONTEXT:some_url]::https::http]
-        [TITLE:xxxxx] Apply .title() to string value. Example: [TITLE:the title]
+
+    - [STRING_WITH_LENGTH_XX] Generates a fixed length string
+    - [INTEGER_WITH_LENGTH_XX] Generates a fixed length integer
+    - [STRING_ARRAY_WITH_LENGTH_XX] Generates a fixed length array of strings
+    - [INTEGER_ARRAY_WITH_LENGTH_XX] Generates a fixed length array of integers
+    - [JSON_WITH_LENGTH_XX] Generates a fixed length JSON
+    - [MISSING_PARAM] Generates a None object
+    - [NULL] Generates a None object
+    - [TRUE] Generates a boolean True
+    - [FALSE] Generates a boolean False
+    - [EMPTY] Generates an empty string
+    - [B] Generates a blank space
+    - [UUID] Generates a v4 UUID
+    - [RANDOM] Generates a random value
+    - [RANDOM_PHONE_NUMBER] Generates a random phone number for language and country configured
+      in dataset.language and dataset.country
+    - [TIMESTAMP] Generates a timestamp from the current time
+    - [DATETIME] Generates a datetime from the current time (UTC)
+    - [NOW] Similar to DATETIME without microseconds; the format depends on the language
+    - [NOW(%Y-%m-%dT%H:%M:%SZ)] Same as NOW but using an specific format by the python strftime function of
+      the datetime module. In the case of the %f placeholder, the syntax has been extended to allow setting
+      an arbitrary number of digits (e.g. %3f would leave just the 3 most significant digits and truncate the rest)
+    - [NOW + 2 DAYS] Similar to NOW but two days later
+    - [NOW - 1 MINUTES] Similar to NOW but one minute earlier
+    - [NOW(%Y-%m-%dT%H:%M:%SZ) - 7 DAYS] Similar to NOW but seven days before and with the indicated format
+    - [TODAY] Similar to NOW without time; the format depends on the language
+    - [TODAY + 2 DAYS] Similar to NOW, but two days later
+    - [ROUND:xxxx::y] Generates a string from a float number (xxxx) with the indicated number of decimals (y)
+    - [STR:xxxx] Cast xxxx to a string
+    - [INT:xxxx] Cast xxxx to an int
+    - [FLOAT:xxxx] Cast xxxx to a float
+    - [LIST:xxxx] Cast xxxx to a list
+    - [DICT:xxxx] Cast xxxx to a dict
+    - [UPPER:xxxx] Converts xxxx to upper case
+    - [LOWER:xxxx] Converts xxxx to lower case
+    - [REPLACE:xxxxx::yy::zz] Replace elements in string. Example: [REPLACE:[CONTEXT:some_url]::https::http]
+    - [TITLE:xxxxx] Apply .title() to string value. Example: [TITLE:the title]
+
     If infer_param_type is True and the result of the replacement process is a string,
     this function also tries to infer and cast the result to the most appropriate data type,
     attempting first the direct conversion to a Python built-in data type and then,
@@ -262,8 +265,12 @@ def _replace_param_transform_string(param):
             except json.decoder.JSONDecodeError:
                 new_param = eval(type_mapping_match_group.group(2))
         elif type_mapping_match_group.group(1) in ['INT', 'FLOAT']:
-            exec(f'exec_param = {type_mapping_match_group.group(1).lower()}({type_mapping_match_group.group(2)})')
-            new_param = locals()['exec_param']
+            exec_env = {}
+            type_str = type_mapping_match_group.group(1).lower()
+            value_str = type_mapping_match_group.group(2)
+            exec_code = f"exec_param = {type_str}({value_str})"
+            exec(exec_code, {}, exec_env)
+            new_param = exec_env['exec_param']
         else:
             replace_param = _get_substring_replacement(type_mapping_match_group)
             new_param = new_param.replace(type_mapping_match_group.group(), replace_param)
@@ -296,13 +303,38 @@ def _get_substring_replacement(type_mapping_match_group):
     return replace_param
 
 
+def _get_format_with_number_of_decimals(base, language):
+    """
+    Get the format and the number of decimals from the base string.
+    """
+    def _is_only_date(base):
+        return 'TODAY' in base
+
+    def _default_format(base):
+        date_format = '%d/%m/%Y' if language == 'es' else '%Y/%m/%d'
+        if _is_only_date(base):
+            return date_format
+        return f'{date_format} %H:%M:%S'
+
+    format_matcher = re.search(r'\((.*)\)', base)
+    if format_matcher and len(format_matcher.groups()) == 1:
+        time_format = format_matcher.group(1)
+        decimal_matcher = re.search(r'%(\d+)f', time_format)
+        if decimal_matcher and len(decimal_matcher.groups()) == 1:
+            return time_format.replace(decimal_matcher.group(0), '%f'), int(decimal_matcher.group(1))
+        return time_format, None
+    return _default_format(base), None
+
+
 def _replace_param_date(param, language):
     """
     Transform param value in a date after applying the specified delta.
     E.g. [TODAY - 2 DAYS], [NOW - 10 MINUTES]
     An specific format could be defined in the case of NOW this way: NOW('THEFORMAT')
     where THEFORMAT is any valid format accepted by the python
-    [datetime.strftime](https://docs.python.org/3/library/datetime.html#datetime.date.strftime) function
+    [datetime.strftime](https://docs.python.org/3/library/datetime.html#datetime.date.strftime) function.
+    In the case of the %f placeholder, the syntax has been extended to allow setting an arbitrary number of digits
+    (e.g. %3f would leave just the 3 most significant digits and truncate the rest).
 
     :param param: parameter value
     :param language: language to configure date format for NOW and TODAY
@@ -319,28 +351,16 @@ def _replace_param_date(param, language):
         the_units = units.lower()
         return now + datetime.timedelta(**dict([(the_units, the_amount)]))
 
-    def _is_only_date(base):
-        return 'TODAY' in base
-
-    def _default_format(base):
-        date_format = '%d/%m/%Y' if language == 'es' else '%Y/%m/%d'
-        if _is_only_date(base):
-            return date_format
-        return f'{date_format} %H:%M:%S'
-
-    def _get_format(base):
-        format_matcher = re.match(r'.*\((.*)\).*', base)
-        if format_matcher and len(format_matcher.groups()) == 1:
-            return format_matcher.group(1)
-        return _default_format(base)
-
     matcher = _date_matcher()
     if not matcher:
         return param, False
 
     base, amount, units = list(matcher.groups())
-    format_str = _get_format(base)
+    format_str, number_of_decimals = _get_format_with_number_of_decimals(base, language)
     date = _offset_datetime(amount, units)
+    if number_of_decimals:
+        decimals = f"{date.microsecond / 1_000_000:.{number_of_decimals}f}"[2:]
+        format_str = format_str.replace("%f", decimals)
     return date.strftime(format_str), True
 
 
@@ -437,18 +457,19 @@ def map_one_param(param):
     """
     Analyze the pattern in the given string and find out its transformed value.
     Available tags and replacement values:
-        [CONF:xxxx] Value from the config dict in project_config global variable for the key xxxx (dot notation is used
-        for keys, e.g. key_1.key_2.0.key_3)
-        [LANG:xxxx] String from the texts dict in language_terms global variable for the key xxxx, using the language
-        specified in language global variable (dot notation is used for keys, e.g. button.label)
-        [POE:xxxx] Definition(s) from the POEditor terms list in poeditor_terms global variable for the term xxxx
-        [TOOLIUM:xxxx] Value from the toolium config in toolium_config global variable for the key xxxx (key format is
-        section_option, e.g. Driver_type)
-        [CONTEXT:xxxx] Value from the behave context storage dict in behave_context global variable for the key xxxx, or
-        value of the behave context attribute xxxx, if the former does not exist
-        [ENV:xxxx] Value of the OS environment variable xxxx
-        [FILE:xxxx] String with the content of the file in the path xxxx
-        [BASE64:xxxx] String with the base64 representation of the file content in the path xxxx
+
+    - [CONF:xxxx] Value from the config dict in project_config global variable for the key xxxx (dot notation is used
+      for keys, e.g. key_1.key_2.0.key_3)
+    - [LANG:xxxx] String from the texts dict in language_terms global variable for the key xxxx, using the language
+      specified in language global variable (dot notation is used for keys, e.g. button.label)
+    - [POE:xxxx] Definition(s) from the POEditor terms list in poeditor_terms global variable for the term xxxx
+    - [TOOLIUM:xxxx] Value from the toolium config in toolium_config global variable for the key xxxx (key format is
+      section_option, e.g. Driver_type)
+    - [CONTEXT:xxxx] Value from the behave context storage dict in behave_context global variable for the key xxxx, or
+      value of the behave context attribute xxxx, if the former does not exist
+    - [ENV:xxxx] Value of the OS environment variable xxxx
+    - [FILE:xxxx] String with the content of the file in the path xxxx
+    - [BASE64:xxxx] String with the base64 representation of the file content in the path xxxx
 
     :param param: string parameter
     :return: transformed value or the original string if no transformation could be applied
@@ -525,19 +546,22 @@ def map_json_param(param, config, copy=True):
     """
     Find the value of the given param using it as a key in the given dictionary. Dot notation is used,
     so for example "service.vamps.user" could be used to retrieve the email in the following config example:
-    {
-        "services":{
-            "vamps":{
-                "user": "cyber-sec-user@11paths.com",
-                "password": "MyPassword"
+
+    .. code-block:: json
+
+        {
+            "services":{
+                "vamps":{
+                    "user": "cyber-sec-user@11paths.com",
+                    "password": "MyPassword"
+                }
             }
         }
-    }
 
     :param param: key to be searched (dot notation is used, e.g. "service.vamps.user").
     :param config: configuration dictionary
     :param copy: boolean value to indicate whether to work with a copy of the given dictionary or not,
-    in which case, the dictionary content might be changed by this function (True by default)
+     in which case, the dictionary content might be changed by this function (True by default)
     :return: mapped value
     """
     properties_list = param.split(".")
@@ -630,18 +654,19 @@ def get_value_from_context(param, context):
     If the resolved element at one of the tokens is a list and the next token is a key=value expression, then the
     element in the list that matches the key=value expression is selected, e.g. "list.key=value" returns the element
     in the list "list" that has the value for key attribute. So, for example, if the list is:
-    [
-        {"key": "value1", "attr": "attr1"},
-        {"key": "value2", "attr": "attr2"}
-    ]
+
+    .. code-block:: json
+
+        [
+            {"key": "value1", "attr": "attr1"},
+            {"key": "value2", "attr": "attr2"}
+        ]
+
     then "list.key=value2" returns the second element in the list. Also does "list.'key'='value2'",
     "list.'key'=\"value2\"", "list.\"key\"='value2'" or "list.\"key\"=\"value2\"".
 
     There is not limit in the nested levels of dotted tokens, so a key like a.b.c.d will be tried to be resolved as:
-
-    context.storage['a'].b.c.d
-        or
-    context.a.b.c.d
+    context.storage['a'].b.c.d or context.a.b.c.d
 
     :param param: key to be searched (e.g. "last_request_result" / "last_request.result")
     :param context: behave context
@@ -758,18 +783,20 @@ def get_message_property(param, language_terms, language_key):
     :param language_key: language key
     :return: the message mapped to the given key in the given language
     """
-    key_list = param.split(".")
-    language_terms_aux = deepcopy(language_terms)
+    lang_param, expected_lang = param.split('::', 1) if '::' in param else (param, language_key)
+
+    key_list = lang_param.split(".")
+    lang_terms_aux = deepcopy(language_terms)
     try:
         for key in key_list:
-            language_terms_aux = language_terms_aux[key]
-        logger.debug(f"Mapping language param '{param}' to its configured value '{language_terms_aux[language_key]}'")
+            lang_terms_aux = lang_terms_aux[key]
+        logger.debug(f"Mapping language param '{lang_param}' to its configured value '{lang_terms_aux[language_key]}'")
     except KeyError:
-        msg = f"Mapping chain '{param}' not found in the language properties file"
+        msg = f"Mapping chain '{lang_param}' not found in the language properties file"
         logger.error(msg)
         raise KeyError(msg)
 
-    return language_terms_aux[language_key]
+    return lang_terms_aux[expected_lang]
 
 
 def get_translation_by_poeditor_reference(reference, poeditor_terms):